requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==5.0.1
docker-pycreds==0.4.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.9
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.10
install==1.3.4
numpy==1.19.2
nvidia-ml-py3==7.352.0
pandas==1.1.3
pathtools==0.1.2
Pillow==8.0.0
promise==2.3
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
sentry-sdk==0.19.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
torch==1.4.0
torchvision==0.5.0
urllib3==1.25.10
wandb==0.10.7
watchdog==0.10.3
